:arrow_up: file naming: removed package name pre-pending to file names.
