+++ Gopkg.toml
@@ -31,7 +31,7 @@
 # Framework required dependencies
 [[constraint]]
   name = "github.com/ory/fosite"
-  version = "0.13.0"
+  version = "0.16.0"
 
 [[constraint]]
   name = "github.com/pkg/errors"
