+++ cache/manager.go
