+++ client/manager.go
