+++ request/oauth2_authorize_code_storage_test.go
@@ -1,0 +1,18 @@
+package request_test
+
+import (
+	"testing"
+
+	"github.com/ory/fosite/handler/oauth2"
+	"github.com/stretchr/testify/assert"
+
+	"github.com/matthewhartstonge/storage/request"
+)
+
+func TestRequestMongoManager_ImplementsFositeAuthorizeCodeStorageInterface(t *testing.T) {
+	r := &request.MongoManager{}
+
+	var i interface{} = r
+	_, ok := i.(oauth2.AuthorizeCodeStorage)
+	assert.Equal(t, true, ok, "request.MongoManager does not implement interface oauth2.AuthorizeCodeStorage")
+}
