+++ user/manager.go
